+++ main/data.h
@@ -15,7 +15,7 @@ class Data {
 
 public:
   QCustomPlot *plot = nullptr;
-  Data(int station_coords, int target_coords, int dist_to_target,
+  Data(double station_coords, double target_coords, double dist_to_target,
        std::string atmosphere_model, std::string refraction_model,
        std::string task);
   Data();
